separate s3 code from asset constants
